packages/cli/requirements.txt
click==8.1.7
rich==13.7.1
httpx==0.27.0
PyYAML==6.0.1
GitPython==3.1.43
